requirements.txt
streamlit
pandas
plotly
google-api-python-client
python-dotenv
pymysql
sqlite3